+++ src/std/ntv/object.ts
@@ -10,6 +10,12 @@ export function clone<T>(i: T): T {
 	return iterate(i) as T;
 }
 
+export function create<T extends IndexSignature>(array: T[], i: any = null) {
+	const object = {} as { [K in T]: typeof i };
+	for (const key of array) object[key] = i;
+	return object;
+}
+
 export function entries<T extends object>(o: T) {
 	return Object.entries(o) as Entries<T>;
 }
@@ -44,7 +50,7 @@ export function iterate<T extends object, I = T[keyof T]>(
 }
 
 export function keys<T extends object>(o: T) {
-	return Object.keys(o) as Array<keyof T>;
+	return Object.keys(o) as Array<string & keyof T>;
 }
 
 export function pick<Keys extends readonly string[]>(keys: Narrow<Keys>) {